COC: Clarify "community spaces" (#1)
The Code of Conduct's "Scope" section states that it applies within all
community spaces, but that language may not be clear to all contibutors.

This commit adds examples of the technical spaces where the CoC would
apply such as bug reports, commit messages, etc. It also adds a
conference talk as an example of an offline event where someone may be
acting as an "appointed representative" for the group.